# backend/app/swagger.yaml
swagger: "2.0"
info:
  title: School Management System API
  description: |
    REST API for the School Management System (SMS).

    ## Authentication
    Most endpoints require a JWT Bearer token. Log in via **POST /auth/login** to obtain
    `access_token` and `refresh_token`. Pass the access token as:
    ```
    Authorization: Bearer <access_token>
    ```
    The access token expires in **15 minutes**. Use **POST /auth/refresh** with the refresh token
    to get a new access token (refresh tokens expire in 7 days).

    ## Roles
    | Role | Access |
    |------|--------|
    | `admin` | All endpoints |
    | `teacher` | Students, attendance, grades, leave review |
    | `student` | Own data only |
    | `parent` | Parent portal, leave submissions, notifications |

    ## Response Envelope
    All responses follow:
    ```json
    { "success": true, "data": {}, "message": "...", "errors": null }
    ```

    ## Dev Seed Credentials
    | Role | Email | Password |
    |------|-------|----------|
    | Admin | admin@school.sms | Admin@1234 |
    | Teacher | priya.sharma@school.sms | Teacher@123 |
    | Student | alice.johnson@school.sms | Student@123 |
    | Parent | robert.johnson@parents.school.sms | Parent@123 |

  version: "1.0.0"
  contact:
    email: admin@school.sms
basePath: /api/v1
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "Format: Bearer <JWT access token>"
...
definitions:
...
  User:
    type: object
    properties:
      id:
        type: integer
        example: 1
      email:
        type: string
        example: admin@school.sms
      role:
        type: string
        enum: [admin, teacher, student, parent]
        example: admin
      first_name:
        type: string
        example: Rahul
      last_name:
        type: string
        example: Verma
      full_name:
        type: string
        example: Rahul Verma
      photo_url:
        type: string
        nullable: true
        example: /uploads/profile_photos/user_1_abc123.jpg
      is_active:
        type: boolean
        example: true
      last_login:
        type: string
        format: date-time
        nullable: true
      created_at:
        type: string
        format: date-time
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      description: Authenticate with email and password. Returns JWT access + refresh tokens.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [email, password]
            properties:
              email:
                type: string
                example: admin@school.sms
              password:
                type: string
                example: Admin@1234
      responses:
        200:
          description: Login successful
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              data:
                type: object
                properties:
                  access_token:
                    type: string
                    example: eyJhbGci...
                  refresh_token:
                    type: string
                    example: eyJhbGci...
                  user:
                    $ref: '#/definitions/User'
        400:
          description: Missing email or password
        401:
          description: Invalid credentials or inactive account
        429:
          description: Rate limit exceeded (5 requests/minute)
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: |
        Use the **refresh token** (not the access token) in the Authorization header to get
        a new access token. Rate limited to 10/minute.
      security:
        - Bearer: []
      responses:
        200:
          description: New access token issued
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              data:
                type: object
                properties:
                  access_token:
                    type: string
        401:
          description: Invalid or expired refresh token
  /auth/logout:
    delete:
      tags: [Auth]
      summary: Logout (revoke token)
      description: Adds the current access token's JTI to the revocation blocklist.
      security:
        - Bearer: []
      responses:
        200:
          description: Logged out successfully
        401:
          description: Missing or invalid token
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      description: Returns the authenticated user's profile. Works for all roles.
      security:
        - Bearer: []
      responses:
        200:
          description: User profile
          schema:
            type: object
            properties:
              success:
                type: boolean
              data:
                $ref: '#/definitions/User'
        401:
          description: Unauthorized
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset
      description: |
        Sends a password reset link to the email address. Always returns 200 to prevent
        email enumeration. In dev mode, the link is logged to console.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [email]
            properties:
              email:
                type: string
                example: admin@school.sms
      responses:
        200:
          description: Reset email sent (or silently skipped if email not found)
        400:
          description: Email field missing
        429:
          description: Rate limit exceeded (3/minute)
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password using token
      description: |
        Provide the token from the reset email link plus the new password.
        Password must be 8+ chars with uppercase, lowercase, digit, and special character.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [token, password]
            properties:
              token:
                type: string
                example: abc123xyz...
              password:
                type: string
                example: NewPass@5678
      responses:
        200:
          description: Password reset successfully
        400:
          description: Missing fields or invalid/expired token
        422:
          description: Password does not meet requirements
  /auth/profile:
    patch:
      tags: [Auth]
      summary: Update own profile
      description: Update first_name and/or last_name of the authenticated user.
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              first_name:
                type: string
                example: Rahul
              last_name:
                type: string
                example: Verma
      responses:
        200:
          description: Profile updated
          schema:
            type: object
            properties:
              data:
                $ref: '#/definitions/User'
        400:
          description: No valid fields to update
        401:
          description: Unauthorized
  /auth/profile/photo:
    post:
      tags: [Auth]
      summary: Upload profile photo
      description: Upload a profile photo (PNG, JPG, JPEG, GIF, WEBP). Max 5MB.
      security:
        - Bearer: []
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: photo
          type: file
          required: true
          description: Image file (PNG, JPG, JPEG, GIF, WEBP — max 5MB)
      responses:
        200:
          description: Photo uploaded
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  photo_url:
                    type: string
                    example: /uploads/profile_photos/user_1_abc123.jpg
        400:
          description: No file, invalid extension, or file too large
        401:
          description: Unauthorized

  # ── Users ────────────────────────────────────────────────────────────────────
...